styled the repo table and made some changes to the root styling of the website
