+++ package.json
@@ -17,9 +17,10 @@
 	"version": "0.1.0",
 	"private": true,
 	"dependencies": {
+		"@material-ui/core": "^4.11.3",
+		"bootstrap": "^4.4.1",
 		"framer": "^1.2.5",
 		"framer-motion": "^2.0.0 ",
-		"bootstrap": "^4.4.1",
 		"fund": "^1.0.0",
 		"react": "^16.8.6",
 		"react-bootstrap": "^1.0.0-beta.16",
