+++ src/components/Home.js
@@ -1,18 +1,14 @@
 import React, { Component } from "react";
 import Main from "../components/Main";
 import Contact from "../components/Contact";
-import Navbar from "../components/Navbar";
 
 
 class Home extends Component {
 	render() {
 		return (
 			<>
-
-					<Navbar />
 					<Main repos={this.props.repos} />
-					<Contact />
-
+					<Contact  repos={this.props.repos}/>
 			</>
 		);
 	}
