+++ src/components/table/table.css
@@ -1,0 +1,54 @@
+table {
+    border-collapse: separate;
+    border-spacing: 0 10px;
+}
+
+th {
+    font-size: 1.2em;
+    font-family: 'Libre Franklin', sans-serif;
+    background-color: #5dd39e;
+    color: black;
+    font-weight: 100;
+    text-align: left !important;
+    
+}
+
+th,
+td,
+tr {
+    width: 400px;
+    text-align: center;
+    border: 1px solid #5dd39e;
+    padding: 10px;
+}
+
+.flex-header {
+    width: 600px;
+}
+
+.repo-URL:hover {
+    background-color: #5dd39e;
+    color: black !important;
+}
+
+a:hover {
+    color: white !important;
+}
+
+.table--container::-webkit-scrollbar {
+    width: 3px;
+    line-height: normal; 
+    margin-left: 500px;
+}
+
+.table--container::-webkit-scrollbar-track {
+    box-shadow: inset 0 0 15px rgb(250, 248, 248);
+    
+}
+
+.table--container::-webkit-scrollbar-thumb {
+    background-color: black;
+    border-radius: 3px;
+    margin: 500px;
+    
+}
